#16 Fix execution of synced chores.  There were an issue that dupes refused to start working (fixed by clearing `driver` field). Another issue was that work were consistently restarting, that was caused by reassigning chores all the time.
 Be aware that FindNextChorePatch.cs is disabled since it is WIP.
