+++ MultiplayerMod/Multiplayer/Configuration/ServerEventBindings.cs
@@ -1,7 +1,7 @@
 ﻿using MultiplayerMod.Core.Dependency;
 using MultiplayerMod.Game;
 using MultiplayerMod.Game.Chores;
-using MultiplayerMod.Multiplayer.Commands;
+using MultiplayerMod.Multiplayer.Commands.Chore;
 using MultiplayerMod.Multiplayer.Commands.Debug;
 using MultiplayerMod.Multiplayer.Commands.State;
 using MultiplayerMod.Multiplayer.Debug;
@@ -11,13 +11,15 @@ using MultiplayerMod.Network;
 
 namespace MultiplayerMod.Multiplayer.Configuration;
 
-public class ServerEventBindings {
+public class ServerEventBindings
+{
 
     private readonly Core.Logging.Logger log = new(typeof(ServerEventBindings));
     private readonly IMultiplayerServer server = Container.Get<IMultiplayerServer>();
     private bool bound;
 
-    public void Bind() {
+    public void Bind()
+    {
         if (bound)
             return;
 
@@ -30,22 +32,8 @@ public class ServerEventBindings {
         WorldDebugSnapshotRunner.SnapshotAvailable += snapshot => server.Send(new SyncWorldDebugSnapshot(snapshot));
         SaveLoaderEvents.WorldSaved += WorldManager.Sync;
 
-        ChoreConsumerEvents.FindNextChore += ChoreConsumerOnFindNextChore;
+        ChoreConsumerEvents.FindNextChore += p => server.Send(new FindNextChore(p), MultiplayerCommandOptions.SkipHost);
 
         bound = true;
     }
-
-    private void ChoreConsumerOnFindNextChore(object sender, FindNextChoreEventArgs e) {
-        server.Send(
-            new FindNextChore(
-                sender.GetType().ToString(),
-                e.InstanceId,
-                e.ChoreId,
-                e.ChoreType.ToString(),
-                e.ChoreCell
-            ),
-            MultiplayerCommandOptions.SkipHost
-        );
-    }
-
 }
